TASK: Move form objects from `Eel` to `Domain` namespace and remove the misleading `Helper` postfix
