+++ Resources/Private/Fusion/Prototypes/Component/Option.fusion
@@ -11,7 +11,7 @@ prototype(Neos.Fusion.Form:Component.Option)  < prototype(Neos.Fusion:Component)
     @ignoreProperties = ${['option']}
 
     @propTypes {
-        option = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Eel\\OptionHelper')}
+        option = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Domain\\Option')}
         attributes = ${PropTypes.dataStructure({})}
         content = ${PropTypes.string}
     }
