+++ Resources/Private/Fusion/Prototypes/Definition/Field.fusion
@@ -3,8 +3,8 @@ prototype(Neos.Fusion.Form:Definition.Field) {
     @class = 'Neos\\Fusion\\Form\\FusionObjects\\FieldDefinitionImplementation'
 
     @propTypes {
-        form = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Eel\\FormHelper')}
-        field = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Eel\\FieldHelper')}
+        form = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Domain\\Form')}
+        field = ${PropTypes.instanceOf('\\Neos\\Fusion\\Form\\Domain\\Field')}
         name = ${PropTypes.string}
         value = ${PropTypes.any}
         multiple = ${PropTypes.boolean}
